Transfered base population and properties to base class.
