+++ components/tsp_algorithms/genetic_algorithm.py
@@ -1,5 +1,3 @@
-import random
-
 from components.evaluate.tsp_evaluate import evaluate
 from components.ga_components.selection import BaseSelector
 from components.ga_components.crossover import BaseCrossover
@@ -22,46 +20,25 @@ class GeneticAlgorithmTsp(BaseTspAlgorithm):
         self.verbose = verbose
         self.atomic_bomb = atomic_bomb
 
-    @property
-    def results(self):
-        return self.__results
-
-    @property
-    def paths(self):
-        return self.__paths
-
-    @property
-    def best_result(self):
-        return self.__best_result
-
-    @property
-    def best_path(self):
-        return self.__best_path
-
     def learn(self, cost_matrix: list):
-        self.__results = {}
-        self.__paths = {}
-        self.__best_result = 0
-        self.__best_path = []
+        self._results = {}
+        self._paths = {}
+        self._best_result = 0
+        self._best_path = []
 
         iteration = 0
         population = self._create_base_population(len(cost_matrix))
         evaluated_population = evaluate(population, cost_matrix)
 
-        if not self.maximize:
-            self.__results[iteration] = min(
-                list(evaluated_population.values()))
-            self.__paths[iteration] = population[
-                min(evaluated_population, key=evaluated_population.get)]
-        else:
-            self.__results[iteration] = max(
-                list(evaluated_population.values()))
-            self.__paths[iteration] = population[
-                max(evaluated_population, key=evaluated_population.get)]
+        self._results[iteration] = min(
+            list(evaluated_population.values()))
+        self._paths[iteration] = population[
+            min(evaluated_population, key=evaluated_population.get)]
+
         if self.verbose:
             print('Starting algorithm...')
             print(f'Iteration: {iteration}. \
-                  Best result: {self.__results[iteration]}')
+                  Best result: {self._results[iteration]}')
 
         while not iteration == self.n_iterations:
             # Increment iteration here because iteration 0 was base population
@@ -78,48 +55,28 @@ class GeneticAlgorithmTsp(BaseTspAlgorithm):
             # Evaluation
             evaluated_population = evaluate(population, cost_matrix)
 
-            if not self.maximize:
-                self.__results[iteration] = min(
-                    list(evaluated_population.values()))
-                self.__paths[iteration] = population[
-                    min(evaluated_population, key=evaluated_population.get)]
-            else:
-                self.__results[iteration] = max(
-                    list(evaluated_population.values()))
-                self.__paths[iteration] = population[
-                    max(evaluated_population, key=evaluated_population.get)]
+            self._results[iteration] = min(
+                list(evaluated_population.values()))
+            self._paths[iteration] = population[
+                min(evaluated_population, key=evaluated_population.get)]
+
             if self.verbose and iteration % 100 == 0:
                 print(f'Iteration: {iteration}. \
-                      Best result: {self.__results[iteration]}')
+                      Best result: {self._results[iteration]}')
 
             # Atomic bomb component - kill population after stagnation
             if (self.atomic_bomb is not None
                 and iteration > self.atomic_bomb
-                    and (self.__results[iteration] == self.__results[
+                    and (self._results[iteration] == self._results[
                         iteration - self.atomic_bomb])):
 
                 population = self._create_base_population(len(cost_matrix))
 
-        if not self.maximize:
-            self.__best_result = min(list(self.__results.values()))
-            self.__best_path = self.__paths[min(self.__results,
-                                                key=self.__results.get)]
-        else:
-            self.__best_result = min(list(self.__results.values()))
-            self.__best_path = self.__paths[min(self.__results,
-                                                key=self.__results.get)]
+            self._best_result = min(list(self._results.values()))
+            self._best_path = self._paths[min(self._results,
+                                              key=self._results.get)]
+
         if self.verbose:
             print('Algorithm finished.')
 
         return self
-
-    def _create_base_population(self, city_size: int) -> dict:
-        base_population = {}
-
-        for i in range(self.population_size):
-            random_individual = list(range(city_size))
-            random.shuffle(random_individual)
-
-            base_population[i] = random_individual
-
-        return base_population
